requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2.5
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.10
gunicorn==23.0.0
celery==5.4.0
redis==5.0.8
django-redis==5.4.0
idna==3.10
inflection==0.5.1
Markdown==3.8.2
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
